Fix DecodingError when LRCLIB returns float duration
LRCLIBLyricsResponse decoded duration as Int, but the LRCLIB API
returns floating-point values (e.g. 213.47). JSONDecoder throws a
typeMismatch DecodingError for any non-integer duration, which
LyricsService surfaced directly as "Lyrics Unavailable". Changing
the field to Double makes decoding succeed for both integer and
fractional durations.

+++ Sources/Sonora/Services/LRCLIBProvider.swift
@@ -229,7 +229,7 @@ struct LRCLIBLyricsResponse: Decodable {
   let trackName: String
   let artistName: String
   let albumName: String?
-  let duration: Int
+  let duration: Double
   let plainLyrics: String?
   let syncedLyrics: String?
